SDK-1302: Remove check module
