+++ spec/yoti/doc_scan/session/create/requested_liveness_check_spec.rb
@@ -2,10 +2,10 @@
 
 require 'spec_helper'
 
-describe 'Yoti::DocScan::Session::Create::Check::RequestedLivenessCheck' do
+describe 'Yoti::DocScan::Session::Create::RequestedLivenessCheck' do
   describe '.to_json' do
     it 'serializes the check' do
-      check = Yoti::DocScan::Session::Create::Check::RequestedLivenessCheck
+      check = Yoti::DocScan::Session::Create::RequestedLivenessCheck
               .builder
               .for_zoom_liveness
               .with_max_retries(3)
